+++ cmd/queuecontroller/app/app.go
@@ -1,0 +1,102 @@
+// Copyright 2025 NVIDIA CORPORATION
+// SPDX-License-Identifier: Apache-2.0
+
+package app
+
+import (
+	"flag"
+
+	enginev2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2"
+	kubeAiSchedulerv2alpha2 "github.com/NVIDIA/KAI-scheduler/pkg/apis/scheduling/v2alpha2"
+	"go.uber.org/zap/zapcore"
+
+	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
+	// to ensure that exec-entrypoint and run can make use of them.
+	_ "k8s.io/client-go/plugin/pkg/client/auth"
+
+	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/healthz"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+
+	"github.com/NVIDIA/KAI-scheduler/pkg/queuecontroller/controllers"
+	// +kubebuilder:scaffold:imports
+)
+
+var (
+	scheme   = runtime.NewScheme()
+	setupLog = ctrl.Log.WithName("setup")
+)
+
+// +kubebuilder:webhook:path=/validate--v1-queue,mutating=false,failurePolicy=fail,sideEffects=None,resources=queues.scheduling.run.ai,verbs=create;update,groups=core,versions=v2,name=queuecontroller.run.ai,admissionReviewVersions=v1
+
+func init() {
+	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
+
+	utilruntime.Must(enginev2.AddToScheme(scheme))
+	utilruntime.Must(kubeAiSchedulerv2alpha2.AddToScheme(scheme))
+	// +kubebuilder:scaffold:scheme
+}
+
+func Run() error {
+	var opts Options
+	opts.AddFlags(flag.CommandLine)
+
+	initLogger()
+
+	var err error
+	options := ctrl.Options{
+		Scheme:           scheme,
+		LeaderElection:   opts.EnableLeaderElection,
+		LeaderElectionID: "42ece193.run.ai",
+	}
+
+	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), options)
+	if err != nil {
+		setupLog.Error(err, "unable to start manager")
+		return nil
+	}
+
+	if err = (&enginev2.Queue{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook for queue v2", "webhook", "Queue")
+		return nil
+	}
+
+	if err = (&controllers.QueueReconciler{
+		Client: mgr.GetClient(),
+		Scheme: mgr.GetScheme(),
+	}).SetupWithManager(mgr, opts.SchedulingQueueLabelKey); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Queue")
+		return nil
+	}
+	// +kubebuilder:scaffold:builder
+
+	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
+		setupLog.Error(err, "unable to set up health check")
+		return nil
+	}
+	if err := mgr.AddReadyzCheck("readyz", healthz.Ping); err != nil {
+		setupLog.Error(err, "unable to set up ready check")
+		return err
+	}
+
+	setupLog.Info("starting manager")
+	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
+		setupLog.Error(err, "problem running manager")
+		return err
+	}
+
+	return nil
+}
+
+func initLogger() {
+	logOptions := zap.Options{
+		Development: true,
+		TimeEncoder: zapcore.ISO8601TimeEncoder,
+	}
+	logOptions.BindFlags(flag.CommandLine)
+	flag.Parse()
+	ctrl.SetLogger(zap.New(zap.UseFlagOptions(&logOptions)))
+}