+++ pkg/queuecontroller/common/common.go
@@ -1,0 +1,8 @@
+// Copyright 2025 NVIDIA CORPORATION
+// SPDX-License-Identifier: Apache-2.0
+
+package common
+
+const (
+	ParentQueueIndexName = ".spec.parentQueue"
+)